Add prime generators
Add interface, test case generators, and a sample implementation for prime number generators.
In this approach, multiple methods are used to generate prime numbers with different integer types (Int32, Int64, BigInteger). In the future, only implementations using BigInteger will be used. If an implementation using a different integer type is needed, it should be trivial to manually change the given code to use a different type.